Move api tests near code
