+++ portal/api/test_client.py
@@ -1,10 +1,10 @@
 import requests
 
-from portal.api import ApiClient
+from portal.api.client import ApiClient
 
 
 def test_api_get_entity_types(mocker):
     mocker.patch('requests.get')
-    api = ApiClient('http://FAKE')
-    api.get_entity_types()
+    client = ApiClient('http://FAKE')
+    client.get_entity_types()
     requests.get.assert_called_once_with('http://FAKE/entities')
